[CHEF-3715] rename files checksum_cache -> digester
