+++ lib/chef/cookbook_uploader.rb
@@ -3,7 +3,7 @@ require 'set'
 require 'rest_client'
 require 'chef/exceptions'
 require 'chef/knife/cookbook_metadata'
-require 'chef/checksum_cache'
+require 'chef/digester'
 require 'chef/cookbook_version'
 require 'chef/cookbook/syntax_check'
 require 'chef/cookbook/file_system_file_vendor'
