+++ lib/chef/digester.rb
