+++ spec/unit/digester_spec.rb
